requirements.txt
PyMuPDF==1.18.3
opencv_python==4.4.0.46
datasets==1.1.3
numpy==1.19.2
torch==1.7.0
en_core_web_sm==2.3.1
pandas==1.1.4
transformers.egg==info
lxml==4.6.1
tqdm==4.49.0
spacy==2.3.2
scenedetect==0.5.4.1
dataclasses==0.7
streamlit==0.71.0
nltk==3.5
fitz==0.0.1.dev2
Pillow==8.1.0
scikit_learn==0.24.0
transformers==4.1.1
